CrudSpec and NodeSpec both are looking for the read method in the raw schema we get back from the sever. However in the raw schema this the read field is called get. This meant non of the read/get methods are appearing in the generator correctly as getReadSpec always returned null.
This changes read to get and everything seems to be working correctly
now, however not knowing the SDK well a sanity check on if this would
break anything woudl be best.

Patched in a hack to correct a error introduced by a poorly checked
schema that is generated by the original ads team schemas.